Don't do the htons thing
Getting at htons on Windows required more than we were willing to do, so
we added a byte-swapping thing that should have worked but turned out to have
issues in some cases. (maybe because of the optimizer? not sure).

To fix, don't do byteswapping and memcpy for a measly 2 bytes; instead, just
copy the bytes one at a time.